@@ -25,24 +25,25 @@ classifiers =
 packages = find:
 package_dir =
     amici = amici
-python_requires = >=3.7
+python_requires = >=3.8
 install_requires =
-    sympy>=1.7.1
-    numpy>=1.14.5; python_version=='3.7'
+    sympy>=1.9
     numpy>=1.17.5; python_version=='3.8'
-    numpy>=1.19.3; python_version>='3.9'
+    numpy>=1.19.3; python_version=='3.9'
+    numpy>=1.21.4; python_version>='3.10'
     python-libsbml
     h5py
     pandas
     pkgconfig
     wurlitzer
     toposort
+    setuptools>=48
 include_package_data = True
 zip_safe = False
 
 [options.extras_require]
 petab = petab>=0.1.17
-pysb = pysb>=1.11.0
+pysb = pysb>=1.13.1
 
 [options.package_data]
 amici =
