@@ -137,9 +137,9 @@ class CVodeSolver : public Solver {
 
     void setErrHandlerFn() const override;
 
-    void setUserData(Model *model) const override;
+    void setUserData() const override;
 
-    void setUserDataB(int which, Model *model) const override;
+    void setUserDataB(int which) const override;
 
     void setMaxNumSteps(long int mxsteps) const override;
 
