@@ -1,6 +1,7 @@
 #ifndef _amici_model_dirac_h
 #define _amici_model_dirac_h
-/* Generated by amiwrap (R2017b) 2bd39716102b9842354964196889d40d2eb2731c */
+/* Generated by amiwrap (R2017b) c7c57bb5806197acc55c77da7272972ad833b5ca */
+/* Generated by amiwrap (R2017b) dc99fa1bcb04914008ba8ed39157513d443c3f43 */
 #include <cmath>
 #include <memory>
 #include "amici/defines.h"
@@ -67,7 +68,7 @@ public:
 
     amici::Model* clone() const override { return new Model_model_dirac(*this); };
 
-    std::string getAmiciCommit() const override { return "2bd39716102b9842354964196889d40d2eb2731c"; };
+    std::string getAmiciCommit() const override { return "c7c57bb5806197acc55c77da7272972ad833b5ca"; };
 
     void fJSparse(SUNMatrixContent_Sparse JSparse, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const realtype *w, const realtype *dwdx) override {
         JSparse_model_dirac(JSparse, t, x, p, k, h, w, dwdx);
@@ -103,7 +104,7 @@ public:
     void fdJzdz(double *dJzdz, const int iz, const realtype *p, const realtype *k, const double *z, const double *sigmaz, const double *mz) override {
     }
 
-    void fdeltaqB(double *deltaqB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ip, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB) override {
+    void fdeltaqB(double *deltaqB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ip, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB, const realtype *xBdot) override {
     }
 
     void fdeltasx(double *deltasx, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const realtype *w, const int ip, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *sx, const realtype *stau) override {
@@ -114,7 +115,7 @@ public:
         deltax_model_dirac(deltax, t, x, p, k, h, ie, xdot, xdot_old);
     }
 
-    void fdeltaxB(double *deltaxB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB) override {
+    void fdeltaxB(double *deltaxB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB, const realtype *xBdot) override {
     }
 
     void fdrzdp(double *drzdp, const int ie, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ip) override {
