@@ -1,6 +1,7 @@
 #ifndef _amici_model_events_h
 #define _amici_model_events_h
-/* Generated by amiwrap (R2017b) 2bd39716102b9842354964196889d40d2eb2731c */
+/* Generated by amiwrap (R2017b) c7c57bb5806197acc55c77da7272972ad833b5ca */
+/* Generated by amiwrap (R2017b) dc99fa1bcb04914008ba8ed39157513d443c3f43 */
 #include <cmath>
 #include <memory>
 #include "amici/defines.h"
@@ -81,7 +82,7 @@ public:
 
     amici::Model* clone() const override { return new Model_model_events(*this); };
 
-    std::string getAmiciCommit() const override { return "2bd39716102b9842354964196889d40d2eb2731c"; };
+    std::string getAmiciCommit() const override { return "c7c57bb5806197acc55c77da7272972ad833b5ca"; };
 
     void fJSparse(SUNMatrixContent_Sparse JSparse, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const realtype *w, const realtype *dwdx) override {
         JSparse_model_events(JSparse, t, x, p, k, h, w, dwdx);
@@ -123,7 +124,7 @@ public:
         dJzdz_model_events(dJzdz, iz, p, k, z, sigmaz, mz);
     }
 
-    void fdeltaqB(double *deltaqB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ip, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB) override {
+    void fdeltaqB(double *deltaqB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ip, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB, const realtype *xBdot) override {
     }
 
     void fdeltasx(double *deltasx, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const realtype *w, const int ip, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *sx, const realtype *stau) override {
@@ -133,7 +134,7 @@ public:
     void fdeltax(double *deltax, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ie, const realtype *xdot, const realtype *xdot_old) override {
     }
 
-    void fdeltaxB(double *deltaxB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB) override {
+    void fdeltaxB(double *deltaxB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB, const realtype *xBdot) override {
     }
 
     void fdrzdp(double *drzdp, const int ie, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ip) override {
