@@ -10,7 +10,7 @@ namespace amici {
 
 namespace model_model_neuron{
 
-void deltaxB_model_neuron(double *deltaxB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB) {
+void deltaxB_model_neuron(double *deltaxB, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const int ie, const realtype *xdot, const realtype *xdot_old, const realtype *xB, const realtype *xBdot) {
               switch(ie) { 
               case 0: {
   deltaxB[0] = xB[0];
