@@ -378,7 +378,7 @@ def import_model_sbml(
         model_output_dir: Optional[str] = None,
         verbose: Optional[Union[bool, int]] = True,
         allow_reinit_fixpar_initcond: bool = True,
-        **kwargs) -> None:
+        **kwargs) -> amici.SbmlImporter:
     """
     Create AMICI model from PEtab problem
 
@@ -415,6 +415,9 @@ def import_model_sbml(
     :param kwargs:
         Additional keyword arguments to be passed to
         :meth:`amici.sbml_import.SbmlImporter.sbml2amici`.
+
+    :return:
+        The created :class:`amici.sbml_import.SbmlImporter` instance.
     """
 
     set_log_level(logger, verbose)
@@ -583,6 +586,8 @@ def import_model_sbml(
         noise_distributions=noise_distrs,
         verbose=verbose,
         **kwargs)
+
+    return sbml_importer
 
 
 # for backwards compatibility
