@@ -510,6 +510,11 @@ std::vector<std::string> Model::getStateNames() const {
     return std::vector<std::string>();
 }
 
+std::vector<std::string> Model::getStateNamesSolver() const {
+    return std::vector<std::string>();
+}
+
+
 bool Model::hasFixedParameterNames() const {
     return nk() == 0 || !getFixedParameterNames().empty();
 }
@@ -547,6 +552,10 @@ bool Model::hasStateIds() const {
 }
 
 std::vector<std::string> Model::getStateIds() const {
+    return std::vector<std::string>();
+}
+
+std::vector<std::string> Model::getStateIdsSolver() const {
     return std::vector<std::string>();
 }
 
@@ -600,7 +609,11 @@ std::vector<bool> const &Model::getStateIsNonNegative() const {
 }
 
 void Model::setStateIsNonNegative(std::vector<bool> const &nonNegative) {
-    if (nx_solver != nx_rdata) {
+    auto any_state_non_negative = std::any_of(nonNegative.begin(),
+                                              nonNegative.end(),
+                                              [](bool x) { return x; });
+
+    if (any_state_non_negative && nx_solver != nx_rdata) {
         throw AmiException("Non-negative states are not supported with"
                            " conservation laws enabled");
     }
@@ -610,9 +623,7 @@ void Model::setStateIsNonNegative(std::vector<bool> const &nonNegative) {
                            state_is_non_negative_.size(), nx_rdata);
     }
     state_is_non_negative_ = nonNegative;
-    any_state_non_negative_ =
-        std::any_of(state_is_non_negative_.begin(), state_is_non_negative_.end(),
-                    [](bool x) { return x; });
+    any_state_non_negative_ = any_state_non_negative;
 }
 
 void Model::setAllStatesNonNegative() {
@@ -779,9 +790,10 @@ void Model::requireSensitivitiesForAllParameters() {
     initializeVectors();
 }
 
-void Model::getExpression(gsl::span<realtype> w, const realtype t, const AmiVector &x)
+void Model::getExpression(gsl::span<realtype> w, const realtype t,
+                          const AmiVector &x)
 {
-    fw(t, x.data());
+    fw(t, computeX_pos(x));
     writeSlice(derived_state_.w_, w);
 }
 
@@ -789,6 +801,10 @@ void Model::getObservable(gsl::span<realtype> y, const realtype t,
                           const AmiVector &x) {
     fy(t, x);
     writeSlice(derived_state_.y_, y);
+}
+
+ObservableScaling Model::getObservableScaling(int /*iy*/) const {
+    return ObservableScaling::lin;
 }
 
 void Model::getObservableSensitivity(gsl::span<realtype> sy, const realtype t,
@@ -907,9 +923,9 @@ void Model::getEventSensitivity(gsl::span<realtype> sz, const int ie,
                                 const realtype t, const AmiVector &x,
                                 const AmiVectorArray &sx) {
     for (int ip = 0; ip < nplist(); ip++) {
-        fsz(&sz.at(ip * nz), ie, t, x.data(), state_.unscaledParameters.data(),
-            state_.fixedParameters.data(), state_.h.data(), sx.data(ip),
-            plist(ip));
+        fsz(&sz[ip * nz], ie, t, computeX_pos(x),
+            state_.unscaledParameters.data(), state_.fixedParameters.data(),
+            state_.h.data(), sx.data(ip), plist(ip));
     }
 }
 
@@ -920,7 +936,7 @@ void Model::getUnobservedEventSensitivity(gsl::span<realtype> sz,
     for (int iz = 0; iz < nz; ++iz)
         if (z2event_.at(iz) - 1 == ie)
             for (int ip = 0; ip < nplist(); ++ip)
-                sz.at(ip * nz + iz) = 0.0;
+                sz[ip * nz + iz] = 0.0;
 }
 
 void Model::getEventRegularization(gsl::span<realtype> rz, const int ie,
@@ -934,8 +950,9 @@ void Model::getEventRegularizationSensitivity(gsl::span<realtype> srz,
                                               const AmiVector &x,
                                               const AmiVectorArray &sx) {
     for (int ip = 0; ip < nplist(); ip++) {
-        fsrz(&srz.at(ip * nz), ie, t, x.data(), state_.unscaledParameters.data(),
-             state_.fixedParameters.data(), state_.h.data(), sx.data(ip),
+        fsrz(&srz[ip * nz], ie, t, computeX_pos(x),
+             state_.unscaledParameters.data(), state_.fixedParameters.data(),
+             state_.h.data(), sx.data(ip),
              plist(ip));
     }
 }
@@ -1057,8 +1074,9 @@ void Model::getEventTimeSensitivity(std::vector<realtype> &stau,
     std::fill(stau.begin(), stau.end(), 0.0);
 
     for (int ip = 0; ip < nplist(); ip++) {
-        fstau(&stau.at(ip), t, x.data(), state_.unscaledParameters.data(),
-              state_.fixedParameters.data(), state_.h.data(), sx.data(ip),
+        fstau(&stau.at(ip), t, computeX_pos(x),
+              state_.unscaledParameters.data(), state_.fixedParameters.data(),
+              state_.h.data(), sx.data(ip),
               plist(ip), ie);
     }
 }
@@ -1066,20 +1084,23 @@ void Model::getEventTimeSensitivity(std::vector<realtype> &stau,
 void Model::addStateEventUpdate(AmiVector &x, const int ie, const realtype t,
                                 const AmiVector &xdot,
                                 const AmiVector &xdot_old) {
-
+    
     derived_state_.deltax_.assign(nx_solver, 0.0);
+    
+    std::copy_n(computeX_pos(x), nx_solver, x.data());
 
     // compute update
-    fdeltax(derived_state_.deltax_.data(), t, x.data(), state_.unscaledParameters.data(),
-            state_.fixedParameters.data(), state_.h.data(), ie, xdot.data(),
-            xdot_old.data());
+    fdeltax(derived_state_.deltax_.data(), t, x.data(),
+            state_.unscaledParameters.data(), state_.fixedParameters.data(),
+            state_.h.data(), ie, xdot.data(), xdot_old.data());
 
     if (always_check_finite_) {
         app->checkFinite(derived_state_.deltax_, "deltax");
     }
 
     // update
-    amici_daxpy(nx_solver, 1.0, derived_state_.deltax_.data(), 1, x.data(), 1);
+    amici_daxpy(nx_solver, 1.0, derived_state_.deltax_.data(), 1, x.data(),
+                1);
 }
 
 void Model::addStateSensitivityEventUpdate(AmiVectorArray &sx, const int ie,
@@ -1096,7 +1117,8 @@ void Model::addStateSensitivityEventUpdate(AmiVectorArray &sx, const int ie,
 
         // compute update
         fdeltasx(derived_state_.deltasx_.data(), t, x_old.data(),
-                 state_.unscaledParameters.data(), state_.fixedParameters.data(),
+                 state_.unscaledParameters.data(),
+                 state_.fixedParameters.data(),
                  state_.h.data(), derived_state_.w_.data(), plist(ip), ie,
                  xdot.data(), xdot_old.data(), sx.data(ip), &stau.at(ip));
 
@@ -1118,7 +1140,7 @@ void Model::addAdjointStateEventUpdate(AmiVector &xB, const int ie,
     derived_state_.deltaxB_.assign(nx_solver, 0.0);
 
     // compute update
-    fdeltaxB(derived_state_.deltaxB_.data(), t, x.data(),
+    fdeltaxB(derived_state_.deltaxB_.data(), t, computeX_pos(x),
              state_.unscaledParameters.data(),
              state_.fixedParameters.data(), state_.h.data(), ie, xdot.data(),
              xdot_old.data(), xB.data(), xBdot.data());
@@ -1141,9 +1163,9 @@ void Model::addAdjointQuadratureEventUpdate(
     for (int ip = 0; ip < nplist(); ip++) {
         derived_state_.deltaqB_.assign(nJ, 0.0);
 
-        fdeltaqB(derived_state_.deltaqB_.data(), t, x.data(),
+        fdeltaqB(derived_state_.deltaqB_.data(), t, computeX_pos(x),
                  state_.unscaledParameters.data(),
-                 state_.fixedParameters.data(), state_.h.data(), ip, ie,
+                 state_.fixedParameters.data(), state_.h.data(), plist(ip), ie,
                  xdot.data(), xdot_old.data(), xB.data(), xBdot.data());
 
         for (int iJ = 0; iJ < nJ; ++iJ)
@@ -1165,10 +1187,6 @@ void Model::updateHeavisideB(const int *rootsfound) {
     for (int ie = 0; ie < ne; ie++) {
         state_.h.at(ie) -= rootsfound[ie];
     }
-}
-
-void Model::updateHeavisideB_eventwise(const int *rootsfound, int ie) {
-    state_.h.at(ie) -= rootsfound[ie];
 }
 
 int Model::checkFinite(gsl::span<const realtype> array, const char *fun) const {
@@ -1208,11 +1226,13 @@ void Model::fx0(AmiVector &x) {
 void Model::fx0_fixedParameters(AmiVector &x) {
     if (!getReinitializeFixedParameterInitialStates())
         return;
+    
     /* we transform to the unreduced states x_rdata and then apply
      x0_fixedparameters to (i) enable updates to states that were removed from
      conservation laws and (ii) be able to correctly compute total abundances
      after updating the state variables */
-    fx_rdata(derived_state_.x_rdata_.data(), x.data(), state_.total_cl.data());
+    fx_rdata(derived_state_.x_rdata_.data(), computeX_pos(x),
+             state_.total_cl.data());
     fx0_fixedParameters(derived_state_.x_rdata_.data(),
                         simulation_parameters_.tstart_,
                         state_.unscaledParameters.data(),
@@ -1233,7 +1253,7 @@ void Model::fsx0(AmiVectorArray &sx, const AmiVector &x) {
         std::fill(derived_state_.sx_rdata_.begin(),
                   derived_state_.sx_rdata_.end(), 0.0);
         fsx0(derived_state_.sx_rdata_.data(), simulation_parameters_.tstart_,
-             x.data(), state_.unscaledParameters.data(),
+             computeX_pos(x), state_.unscaledParameters.data(),
              state_.fixedParameters.data(), plist(ip));
         fsx_solver(sx.data(ip), derived_state_.sx_rdata_.data());
         fstotal_cl(stcl, derived_state_.sx_rdata_.data(), plist(ip));
@@ -1249,7 +1269,8 @@ void Model::fsx0_fixedParameters(AmiVectorArray &sx, const AmiVector &x) {
             stcl = &state_.stotal_cl.at(plist(ip) * ncl());
         fsx_rdata(derived_state_.sx_rdata_.data(), sx.data(ip), stcl, plist(ip));
         fsx0_fixedParameters(derived_state_.sx_rdata_.data(),
-                             simulation_parameters_.tstart_, x.data(),
+                             simulation_parameters_.tstart_,
+                             computeX_pos(x),
                              state_.unscaledParameters.data(),
                              state_.fixedParameters.data(),
                              plist(ip),
@@ -1262,7 +1283,7 @@ void Model::fsx0_fixedParameters(AmiVectorArray &sx, const AmiVector &x) {
 void Model::fsdx0() {}
 
 void Model::fx_rdata(AmiVector &x_rdata, const AmiVector &x) {
-    fx_rdata(x_rdata.data(), x.data(), state_.total_cl.data());
+    fx_rdata(x_rdata.data(), computeX_pos(x), state_.total_cl.data());
     if (always_check_finite_)
         checkFinite(x_rdata.getVector(), "x_rdata");
 }
@@ -1282,7 +1303,7 @@ void Model::writeSliceEvent(gsl::span<const realtype> slice,
     checkBufferSize(buffer, z2event_.size());
     for (unsigned izt = 0; izt < z2event_.size(); ++izt)
         if (z2event_.at(izt) - 1 == ie)
-            buffer.at(izt) = slice.at(izt);
+            buffer[izt] = slice[izt];
 }
 
 void Model::writeSensitivitySliceEvent(gsl::span<const realtype> slice,
@@ -1293,7 +1314,7 @@ void Model::writeSensitivitySliceEvent(gsl::span<const realtype> slice,
     for (int ip = 0; ip < nplist(); ++ip)
         for (unsigned izt = 0; izt < z2event_.size(); ++izt)
             if (z2event_.at(izt) - 1 == ie)
-                buffer.at(ip * nztrue + izt) = slice.at(ip * nztrue + izt);
+                buffer[ip * nztrue + izt] = slice[ip * nztrue + izt];
 }
 
 void Model::writeLLHSensitivitySlice(const std::vector<realtype> &dLLhdp,
@@ -1327,12 +1348,14 @@ void Model::initializeVectors() {
 void Model::fy(const realtype t, const AmiVector &x) {
     if (!ny)
         return;
+    
+    auto x_pos = computeX_pos(x);
 
     derived_state_.y_.assign(ny, 0.0);
 
-    fw(t, x.data());
-    fy(derived_state_.y_.data(), t, x.data(), state_.unscaledParameters.data(),
-       state_.fixedParameters.data(),
+    fw(t, x_pos);
+    fy(derived_state_.y_.data(), t, x_pos,
+       state_.unscaledParameters.data(), state_.fixedParameters.data(),
        state_.h.data(), derived_state_.w_.data());
 
     if (always_check_finite_) {
@@ -1343,20 +1366,22 @@ void Model::fy(const realtype t, const AmiVector &x) {
 void Model::fdydp(const realtype t, const AmiVector &x) {
     if (!ny)
         return;
+    
+    auto x_pos = computeX_pos(x);
 
     derived_state_.dydp_.assign(ny * nplist(), 0.0);
-    fw(t, x.data());
-    fdwdp(t, x.data());
+    fw(t, x_pos);
+    fdwdp(t, x_pos);
 
     /* get dydp slice (ny) for current time and parameter */
     for (int ip = 0; ip < nplist(); ip++)
         if (pythonGenerated) {
-            fdydp(&derived_state_.dydp_.at(ip * ny), t, x.data(),
+            fdydp(&derived_state_.dydp_.at(ip * ny), t, x_pos,
                   state_.unscaledParameters.data(),
                   state_.fixedParameters.data(), state_.h.data(), plist(ip),
                   derived_state_.w_.data(), state_.stotal_cl.data());
         } else {
-            fdydp(&derived_state_.dydp_.at(ip * ny), t, x.data(),
+            fdydp(&derived_state_.dydp_.at(ip * ny), t, x_pos,
                   state_.unscaledParameters.data(),
                   state_.fixedParameters.data(), state_.h.data(), plist(ip),
                   derived_state_.w_.data(), derived_state_.dwdp_.data());
@@ -1370,14 +1395,17 @@ void Model::fdydp(const realtype t, const AmiVector &x) {
 void Model::fdydx(const realtype t, const AmiVector &x) {
     if (!ny)
         return;
+    
+    auto x_pos = computeX_pos(x);
 
     derived_state_.dydx_.assign(ny * nx_solver, 0.0);
 
-    fw(t, x.data());
-    fdwdx(t, x.data());
-    fdydx(derived_state_.dydx_.data(), t, x.data(), state_.unscaledParameters.data(),
-          state_.fixedParameters.data(), state_.h.data(),
-          derived_state_.w_.data(), derived_state_.dwdx_.data());
+    fw(t, x_pos);
+    fdwdx(t, x_pos);
+    fdydx(derived_state_.dydx_.data(), t, x_pos,
+          state_.unscaledParameters.data(), state_.fixedParameters.data(),
+          state_.h.data(), derived_state_.w_.data(),
+          derived_state_.dwdx_.data());
 
     if (always_check_finite_) {
         app->checkFinite(derived_state_.dydx_, "dydx");
@@ -1603,8 +1631,9 @@ void Model::fz(const int ie, const realtype t, const AmiVector &x) {
 
     derived_state_.z_.assign(nz, 0.0);
 
-    fz(derived_state_.z_.data(), ie, t, x.data(), state_.unscaledParameters.data(),
-       state_.fixedParameters.data(), state_.h.data());
+    fz(derived_state_.z_.data(), ie, t, computeX_pos(x),
+       state_.unscaledParameters.data(), state_.fixedParameters.data(),
+       state_.h.data());
 }
 
 void Model::fdzdp(const int ie, const realtype t, const AmiVector &x) {
@@ -1614,7 +1643,7 @@ void Model::fdzdp(const int ie, const realtype t, const AmiVector &x) {
     derived_state_.dzdp_.assign(nz * nplist(), 0.0);
 
     for (int ip = 0; ip < nplist(); ip++) {
-        fdzdp(derived_state_.dzdp_.data(), ie, t, x.data(),
+        fdzdp(derived_state_.dzdp_.data(), ie, t, computeX_pos(x),
               state_.unscaledParameters.data(),
               state_.fixedParameters.data(), state_.h.data(), plist(ip));
     }
@@ -1630,8 +1659,9 @@ void Model::fdzdx(const int ie, const realtype t, const AmiVector &x) {
 
     derived_state_.dzdx_.assign(nz * nx_solver, 0.0);
 
-    fdzdx(derived_state_.dzdx_.data(), ie, t, x.data(), state_.unscaledParameters.data(),
-          state_.fixedParameters.data(), state_.h.data());
+    fdzdx(derived_state_.dzdx_.data(), ie, t, computeX_pos(x),
+          state_.unscaledParameters.data(), state_.fixedParameters.data(),
+          state_.h.data());
 
     if (always_check_finite_) {
         app->checkFinite(derived_state_.dzdx_, "dzdx");
@@ -1642,7 +1672,7 @@ void Model::frz(const int ie, const realtype t, const AmiVector &x) {
 
     derived_state_.rz_.assign(nz, 0.0);
 
-    frz(derived_state_.rz_.data(), ie, t, x.data(),
+    frz(derived_state_.rz_.data(), ie, t, computeX_pos(x),
         state_.unscaledParameters.data(),
         state_.fixedParameters.data(), state_.h.data());
 }
@@ -1654,7 +1684,7 @@ void Model::fdrzdp(const int ie, const realtype t, const AmiVector &x) {
     derived_state_.drzdp_.assign(nz * nplist(), 0.0);
 
     for (int ip = 0; ip < nplist(); ip++) {
-        fdrzdp(derived_state_.drzdp_.data(), ie, t, x.data(),
+        fdrzdp(derived_state_.drzdp_.data(), ie, t, computeX_pos(x),
                state_.unscaledParameters.data(),
                state_.fixedParameters.data(), state_.h.data(), plist(ip));
     }
@@ -1670,8 +1700,9 @@ void Model::fdrzdx(const int ie, const realtype t, const AmiVector &x) {
 
     derived_state_.drzdx_.assign(nz * nx_solver, 0.0);
 
-    fdrzdx(derived_state_.drzdx_.data(), ie, t, x.data(), state_.unscaledParameters.data(),
-           state_.fixedParameters.data(), state_.h.data());
+    fdrzdx(derived_state_.drzdx_.data(), ie, t, computeX_pos(x),
+           state_.unscaledParameters.data(), state_.fixedParameters.data(),
+           state_.h.data());
 
     if (always_check_finite_) {
         app->checkFinite(derived_state_.drzdx_, "drzdx");
@@ -2092,6 +2123,14 @@ const_N_Vector Model::computeX_pos(const_N_Vector x) {
     }
 
     return x;
+}
+
+const realtype *Model::computeX_pos(AmiVector const& x) {
+    if (any_state_non_negative_) {
+        computeX_pos(x.getNVector());
+        return derived_state_.x_pos_tmp_.data();
+    }
+    return x.data();
 }
 
 void Model::setReinitializationStateIdxs(std::vector<int> const& idxs)
